+++ samples/tfm-comm/tfm-comm3/dynarr-void.h
@@ -2,7 +2,7 @@
 #include <stddef.h>
 
 typedef struct {
-    size_t len;
+    uint64_t len;
     size_t capacity;
     size_t elem_size;
     void* data;
@@ -20,11 +20,11 @@ void dynarr_append(const size_t elem_size, const void* elem, DynamicArray* arr);
 void dynarr_append_multiple(const size_t count, const size_t elem_size, const void* elems,
                             DynamicArray* arr);
 
-void dynarr_remove(const size_t index, DynamicArray* arr);
+void dynarr_remove(const uint64_t index, DynamicArray* arr);
 
-void dynarr_remove_multiple(const size_t index, const size_t count, DynamicArray* arr);
+void dynarr_remove_multiple(const uint64_t index, const size_t count, DynamicArray* arr);
 
-void* dynarr_get(const size_t index, const DynamicArray arr);
+void* dynarr_get(const uint64_t index, const DynamicArray arr);
 
 void print_dynarr(const char* label, const DynamicArray arr);
