+++ samples/tfm-comm/tfm-comm3/segment.h
@@ -1,13 +1,14 @@
 #pragma once
 #include <stddef.h>
+#include <stdint.h>
 
 typedef struct {
     size_t ndims;
-    size_t* dims;
-    size_t* offset;
+    uint64_t* dims;
+    uint64_t* offset;
 } Segment;
 
-Segment segment_create(const size_t ndims, const size_t* dims, const size_t* offset);
+Segment segment_create(const size_t ndims, const uint64_t* dims, const uint64_t* offset);
 
 void segment_destroy(Segment* segment);
