+++ samples/tfm-comm/tfm-comm3/segment_batch.h
@@ -1,13 +1,14 @@
 #pragma once
 #include <stddef.h>
+#include <stdint.h>
 
 #include "packet.h"
 
 struct HaloSegmentBatch_s {
     size_t ndims;
-    size_t* local_mm;
-    size_t* local_nn;
-    size_t* local_nn_offset;
+    uint64_t* local_mm;
+    uint64_t* local_nn;
+    uint64_t* local_nn_offset;
 
     size_t npackets;
     Packet* local_packets;
